+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Formal\Migrations\Failure`
+- `Formal\Migrations\Applied::versions()`
+
 ### Changed
 
 - Requires PHP `8.4`
@@ -9,12 +14,14 @@
 - Requires `innmind/foundation:~2.1`
 - `Formal\Migrations\SQL` has been renamed to `SQL\Runner` and is now flagged as internal
 - `Formal\Migrations\Commands` has been renamed to `Commands\Runner` and is now flagged as internal
-- `Formal\Migrations\Applied` error is now an exception
+- `Formal\Migrations\Factory\SQL::migrate()` now returns an `Innmind\Immutable\Either<Formal\Migrations\Failure, Formal\Migrations\Applied>`
+- `Formal\Migrations\Factory\Commands::migrate()` now returns an `Innmind\Immutable\Either<Formal\Migrations\Failure, Formal\Migrations\Applied>`
 
 ### Removed
 
 - `Formal\Migrations\Runner`
 - `Formal\Migrations\Migration`
+- `Formal\Migrations\Applied::match()`
 
 ## 1.1.0 - 2025-02-28